+++ custom-bulkquick-edit.php
@@ -295,7 +295,7 @@ EOD;
 
 		$fields = self::get_enabled_fields( $post->post_type );
 		foreach ( $fields as $key => $field_name ) {
-			if ( false !== strstr( $field_name, Custom_Bulkquick_Edit_Settings::AUTO ) || false !== strstr( $field_name, Custom_Bulkquick_Edit_Settings::RESET ) )
+			if ( false !== strstr( $field_name, Custom_Bulkquick_Edit_Settings::AUTO ) || false !== strstr( $field_name, Custom_Bulkquick_Edit_Settings::FORCE ) || false !== strstr( $field_name, Custom_Bulkquick_Edit_Settings::RESET ) )
 				continue;
 
 			$title                  = Custom_Bulkquick_Edit_Settings::$settings[ $key ]['label'];
@@ -457,6 +457,12 @@ jQuery(document).ready(function($) {
 			return;
 		}
 
+		if ( false !== strstr( $field_name, Custom_Bulkquick_Edit_Settings::FORCE ) ) {
+			$field_name = str_replace( Custom_Bulkquick_Edit_Settings::FORCE, '', $field_name );
+			delete_post_meta( $post_id, $field_name );
+			return;
+		}
+
 		$value = stripslashes_deep( $value );
 		if ( 'taxonomy' == $field_type ) {
 			// WordPress doesn't keep " enclosed CSV terms together, so
@@ -467,7 +473,7 @@ jQuery(document).ready(function($) {
 		} elseif ( 'categories' == $field_type ) {
 			$value = array_map( 'intval', $value );
 			$value = array_unique( $value );
-			if ( isset( $value[ 0 ] ) && 0 === $value[ 0 ] ) 
+			if ( isset( $value[ 0 ] ) && 0 === $value[ 0 ] )
 				unset( $value[ 0 ] );
 
 			wp_set_object_terms( $post_id, $value, $field_name );
@@ -635,14 +641,8 @@ jQuery(document).ready(function($) {
 			return;
 		}
 
-		if ( 'post_excerpt' == $column_name ) {
+		if ( 'post_excerpt' == $column_name )
 			$field_type = 'textarea';
-		} elseif ( false !== strstr( $column_name, Custom_Bulkquick_Edit_Settings::RESET ) ) {
-			if ( ! $bulk_mode )
-				return;
-			else
-				$field_type = 'checkbox';
-		}
 
 		if ( self::$no_instance ) {
 			self::$no_instance = false;
@@ -713,6 +713,31 @@ jQuery(document).ready(function($) {
 			if ( empty( $do_pre_title ) ) {
 				$result .= $check_title;
 				$result .= '</label>';
+			}
+			if ( $bulk_mode && 'checkbox' == $field_type ) {
+				$key_force = $key . Custom_Bulkquick_Edit_Settings::FORCE;
+				$enable    = cbqe_get_option( $key_force );
+
+				if ( $enable ) {
+					$field_force  = $field_name . Custom_Bulkquick_Edit_Settings::FORCE;
+					$title        = Custom_Bulkquick_Edit_Settings::$settings[ $key_force ]['label'];
+					$column_force = $column_name . Custom_Bulkquick_Edit_Settings::RESET;
+
+					$result  = '';
+					$result .= '<label class="alignleft">';
+					$result .= '<input type="checkbox" name="' . $field_force . '" />';
+					$result .= ' ';
+					$result .= '<span class="checkbox-title">' . $title . '</span>';
+					$result .= '</label>';
+
+					echo $open_fieldset;
+					echo $result;
+					echo $close_fieldset;
+
+					self::$scripts_bulk[ $column_force ] = "'{$field_force}': bulk_row.find( 'input[name^={$field_force}]:checkbox:checked' ).map(function(){ return $(this).val(); }).get()";
+				}
+
+				return;
 			}
 			break;